src/python/requirements.txt
sqlalchemy
pandas
numpy
psycopg2 
pyarrow
requests
python-dotenv